requirements.txt
fastapi
uvicorn
gunicorn
pandas
openpyxl
xlrd
python-multipart
python-dotenv
